+++ wallet-connect-test-bench/front-end/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+
 - Add a new section on the UI to sponsor the transaction invoking a smart contract
 
 ## 1.5.5